backend/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.51.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-quill-editor==0.1.22
django-resized==1.0.2
django-searchable-select==1.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
# djangorestframework-simplejwt==5.2.0
djoser==2.1.0
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.4
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.4
whitenoise==5.2.0
